requirements.txt
docopt-ng==0.9.0
pygame-ce==2.5.3
